fullstack-todo/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Todo API Backend with Authentication"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.104.1",
    "sqlmodel==0.0.16",
    "pydantic[email]>=2.5.0,<2.9.0",
    "pydantic-settings>=2.0.3",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "pytest==7.4.3",
    "httpx==0.25.2",
    "uvicorn[standard]==0.24.0",
    "alembic==1.13.1",
    "sqlalchemy>=2.0.25",
    "python-dotenv==1.0.0",
    "psycopg2-binary>=2.9.7",
    "slowapi==0.1.9",
    "python-multipart==0.0.6",
]
